+++ era-compiler-solidity/src/process/input_evm/dependency_data.rs
@@ -36,16 +36,6 @@ impl DependencyData {
 }
 
 impl era_compiler_llvm_context::Dependency for DependencyData {
-    fn get(&self, identifier: &str) -> anyhow::Result<String> {
-        let path = self.resolve_path(identifier)?;
-        let _contract = self
-            .dependencies
-            .get(path.as_str())
-            .cloned()
-            .ok_or_else(|| anyhow::anyhow!("dependency `{path}` not found in the project"))?;
-        unimplemented!()
-    }
-
     fn resolve_path(&self, identifier: &str) -> anyhow::Result<String> {
         self.identifier_paths
             .get(identifier.strip_suffix("_deployed").unwrap_or(identifier))