+++ era-compiler-solidity/tests/common/mod.rs
@@ -156,16 +156,18 @@ pub fn build_solidity(
     let build = project.compile_to_eravm(
         &mut vec![],
         true,
-        linker_symbols,
         metadata_hash_type,
         optimizer_settings,
         vec![],
         false,
         None,
-        None,
     )?;
-    build.write_to_standard_json(&mut solc_output, Some(&solc_compiler.version))?;
+    build.collect_errors()?;
 
+    let build = build.link(linker_symbols);
+    build.collect_errors()?;
+
+    build.write_to_standard_json(&mut solc_output, Some(&solc_compiler.version))?;
     solc_output.collect_errors()?;
     Ok(solc_output)
 }
@@ -254,16 +256,18 @@ pub fn build_yul(
     let build = project.compile_to_eravm(
         &mut vec![],
         true,
-        BTreeMap::new(),
         era_compiler_common::HashType::Ipfs,
         optimizer_settings,
         vec![],
         false,
         None,
-        None,
     )?;
-    build.write_to_standard_json(&mut solc_output, None)?;
+    build.collect_errors()?;
 
+    let build = build.link(BTreeMap::new());
+    build.collect_errors()?;
+
+    build.write_to_standard_json(&mut solc_output, None)?;
     solc_output.collect_errors()?;
     Ok(solc_output)
 }
@@ -307,16 +311,18 @@ pub fn build_yul_standard_json(
     let build = project.compile_to_eravm(
         &mut vec![],
         solc_compiler.is_none(),
-        BTreeMap::new(),
         era_compiler_common::HashType::Ipfs,
         optimizer_settings,
         vec![],
         false,
         None,
-        None,
     )?;
-    build.write_to_standard_json(&mut solc_output, solc_version)?;
+    build.collect_errors()?;
 
+    let build = build.link(BTreeMap::new());
+    build.collect_errors()?;
+
+    build.write_to_standard_json(&mut solc_output, solc_version)?;
     solc_output.collect_errors()?;
     Ok(solc_output)
 }
@@ -344,16 +350,18 @@ pub fn build_llvm_ir_standard_json(
     let build = project.compile_to_eravm(
         &mut vec![],
         true,
-        BTreeMap::new(),
         era_compiler_common::HashType::Ipfs,
         optimizer_settings,
         vec![],
         false,
         None,
-        None,
     )?;
-    build.write_to_standard_json(&mut output, None)?;
+    build.collect_errors()?;
 
+    let build = build.link(BTreeMap::new());
+    build.collect_errors()?;
+
+    build.write_to_standard_json(&mut output, None)?;
     output.collect_errors()?;
     Ok(output)
 }
@@ -377,16 +385,18 @@ pub fn build_eravm_assembly_standard_json(
     let build = project.compile_to_eravm(
         &mut vec![],
         true,
-        BTreeMap::new(),
         era_compiler_common::HashType::Ipfs,
         optimizer_settings,
         vec![],
         false,
         None,
-        None,
     )?;
-    build.write_to_standard_json(&mut output, None)?;
+    build.collect_errors()?;
 
+    let build = build.link(BTreeMap::new());
+    build.collect_errors()?;
+
+    build.write_to_standard_json(&mut output, None)?;
     output.collect_errors()?;
     Ok(output)
 }